catalyst_agent/catalyst_agent/llm_gemini.py
import os
import google.generativeai as genai
from .llm_base import BaseLLM # Corrected import name
# LLMResponse seems not defined in llm_base, removing for now. Need to define it or import from elsewhere if used.
import os # Import os for environment variables
import google.generativeai as genai # Ensure genai is imported
from google.generativeai.types import GenerationConfig # For parameter mapping
from .utils import setup_logger # Import setup_logger
import logging # Import logging

# Environment variables expected:
# GOOGLE_LLM_API_KEY: Your Google API key for Gemini.
# GEMINI_MODEL_NAME: The specific Gemini model to use (e.g., "gemini-pro").

class GeminiLLM(BaseLLM): # Inherit from the correct base class
    """
    LLM client implementation for Google Gemini models.
    """
    def __init__(self, config=None, logger=None): # Match BaseLLM constructor signature (optional for now)
        # Provide a default logger if none is passed, similar to how LLMManager might handle it
        self.logger = logger or setup_logger('agentic.llm.gemini', logging.INFO)
        # super().__init__(config, self.logger) # Call BaseLLM init if needed, requires config/logger passing
        self.api_key = os.getenv("GOOGLE_LLM_API_KEY") # Use os.getenv
        # Store the model name used during initialization
        self._model_name = os.getenv("GEMINI_MODEL_NAME", "gemini-pro") # Use os.getenv with default
        if not self.api_key:
            raise ValueError("GOOGLE_LLM_API_KEY environment variable not set.")

        try:
            genai.configure(api_key=self.api_key)
            self.model = genai.GenerativeModel(self._model_name) # Use stored model name
        except Exception as e:
            self.logger.error(f"Failed to configure Gemini SDK or get model: {e}", exc_info=True)
            raise ValueError("Gemini LLM initialization failed.") from e
        self.logger.info(f"Initialized GeminiLLM with model: {self._model_name}") # Use stored model name

    # TODO: Adapt this method to match the expected signature and return type of BaseLLM.chat_completion
    # Current BaseLLM expects chat_completion, not a simple generate.
    # Also, LLMResponse is not defined in llm_base.py
    async def generate(self, prompt: str, **kwargs): # -> LLMResponse:
        """
        Generates text using the configured Gemini model. (Needs adaptation for BaseLLM compatibility)

        Args:
            prompt: The input prompt for the model.
            **kwargs: Additional keyword arguments.

        Returns:
            Response text (or adapt to match BaseLLM.chat_completion structure).
        """
        self.logger.debug(f"Generating text with Gemini model {self.model_name}")
        try:
            # Basic generation - adapt parameters as needed based on genai library capabilities
            # Example: generation_config = genai.types.GenerationConfig(temperature=0.7, max_output_tokens=kwargs.get('max_tokens'))
            # response = await self.model.generate_content_async(prompt, generation_config=generation_config)

            # Simple generation for now
            response = await self.model.generate_content_async(prompt)

            # Handle potential streaming or different response structures if needed
            # Assuming response.text gives the desired output for now
            generated_text = response.text

            self.logger.debug(f"Gemini generation successful. Response length: {len(generated_text)}")
            # Returning raw text for now. Needs to be structured like BaseLLM.chat_completion expects.
            # Example structure to eventually return:
            # return {
            #     "choices": [{"message": {"role": "assistant", "content": generated_text}}]
            # }
            return generated_text # Placeholder return

        except Exception as e:
            self.logger.error(f"Error during Gemini generation: {e}", exc_info=True)
            # Consider more specific error handling based on google.generativeai exceptions
            # Returning error string for now. Needs structure.
            # Example structure:
            # return {"error": str(e)}
            return f"Error: {str(e)}" # Placeholder return

    # BaseLLM requires estimate_tokens and model_name property
    def estimate_tokens(self, text: str) -> int:
        """
        Estimate the number of tokens for Gemini.
        NOTE: This is a placeholder. The actual tokenization might differ.
        Google's API might provide a specific way to count tokens.
        """
        # Simple estimation: characters / 4 (very rough)
        estimated_tokens = len(text) // 4
        self.logger.debug(f"Estimated tokens for text (length {len(text)}): {estimated_tokens}")
        return estimated_tokens

    @property
    def model_name(self) -> str:
        """Return the configured Gemini model name."""
        # Return the model name stored during initialization
        return self._model_name

    def chat_completion(self, messages: list[dict[str, str]], temperature: float, max_tokens: int, response_format: dict[str, str] | None = None) -> dict[str, any]:
        """
        Generate a chat completion using the Gemini model, adhering to the BaseLLM interface.
        """
        self.logger.debug(f"Initiating Gemini chat completion with {len(messages)} messages.")

        # --- Parameter Mapping ---
        # Map BaseLLM parameters to Gemini's GenerationConfig
        # Note: Gemini might use different names or have limitations (e.g., max_output_tokens)
        generation_config_args = {
            "temperature": temperature,
            "max_output_tokens": max_tokens
            # TODO: Add mapping for other potential parameters if needed (top_p, top_k, etc.)
        }

        # --- JSON Mode Handling ---
        # Check if JSON output is requested
        if response_format and response_format.get("type") == "json_object":
            self.logger.info("JSON output requested, enabling Gemini JSON mode.")
            # Enable Gemini's JSON mode via response_mime_type
            generation_config_args["response_mime_type"] = "application/json"
            # Ensure the prompt explicitly asks for JSON (LLMManager should already do this, but good practice)
            # Example: Append to system prompt if needed: "Ensure your response is a valid JSON object."

        generation_config = GenerationConfig(**generation_config_args)
        self.logger.debug(f"Using GenerationConfig: {generation_config}")

        # --- Message Formatting ---
        # Convert the message list to a format suitable for Gemini.
        # Gemini's `generate_content` often works best with alternating user/model roles.
        # We might need to condense the history or handle system prompts carefully.
        # For simplicity, let's concatenate for now, but this might need refinement.
        # A better approach might involve passing the structured history if the model supports it.
        prompt_history = []
        system_prompt = ""
        for msg in messages:
            role = msg.get("role")
            content = msg.get("content")
            if role == "system":
                # System prompts might need special handling or prepending.
                # For now, let's store it separately.
                system_prompt += content + "\n"
            elif role == "user":
                prompt_history.append({"role": "user", "parts": [content]})
            elif role == "assistant":
                 # Gemini uses 'model' role for assistant messages
                prompt_history.append({"role": "model", "parts": [content]})
            else:
                self.logger.warning(f"Unsupported role '{role}' in message history, skipping.")

        # Prepend system prompt to the first user message if applicable
        if system_prompt and prompt_history and prompt_history[0]["role"] == "user":
             prompt_history[0]["parts"][0] = system_prompt + prompt_history[0]["parts"][0]
        elif system_prompt:
             # If no user message, add system prompt as initial user message (may need adjustment)
             prompt_history.insert(0, {"role": "user", "parts": [system_prompt]})

        # If history is empty after processing, create a default prompt
        if not prompt_history:
             prompt_history = [{"role": "user", "parts": ["Hello."]}] # Default if no messages
             self.logger.warning("Message history was empty or only contained system prompts. Using default 'Hello.'")


        self.logger.debug(f"Formatted prompt history for Gemini: {prompt_history}")

        # --- API Call ---
        try:
            # Use the synchronous generate_content method
            response = self.model.generate_content(
                contents=prompt_history,
                generation_config=generation_config
                # TODO: Handle response_format if Gemini supports JSON mode directly
            )
            self.logger.debug(f"Raw Gemini response received: {response}")

            # --- Response Formatting ---
            # Extract the generated text and format it according to BaseLLM's expected structure.
            # Need to handle potential errors or empty responses from Gemini.
            # Check candidates first as response.parts requires a candidate
            if response.candidates and response.candidates[0].content.parts:
                 # Sometimes the content is nested under candidates
                 generated_text = "".join(part.text for part in response.candidates[0].content.parts)
            elif response.parts: # Check this second, as per original logic
                 generated_text = "".join(part.text for part in response.parts)
            else:
                 # Handle cases where no text is generated (e.g., safety filters or empty response)
                 generated_text = ""
                 # Log safety feedback if available
                 if response.prompt_feedback:
                     self.logger.warning(f"Gemini prompt feedback: {response.prompt_feedback}")
                 if response.candidates and response.candidates[0].finish_reason != "STOP":
                     self.logger.warning(f"Gemini finish reason: {response.candidates[0].finish_reason}")
                     generated_text = f"[Generation stopped: {response.candidates[0].finish_reason}]"


            self.logger.info(f"Gemini chat completion successful. Response length: {len(generated_text)}")

            # Format response like OpenAI structure expected by BaseLLM
            return {
                "choices": [
                    {
                        "message": {
                            "role": "assistant",
                            "content": generated_text
                        },
                        "finish_reason": response.candidates[0].finish_reason.name if response.candidates else "UNKNOWN"
                        # Add other relevant fields if needed
                    }
                ],
                "model": self._model_name, # Include model name used
                "usage": { # Placeholder for token usage - Gemini API might provide this
                    "prompt_tokens": None,
                    "completion_tokens": None,
                    "total_tokens": None
                }
            }

        except Exception as e:
            self.logger.error(f"Error during Gemini chat completion: {e}", exc_info=True)
            # Return an error structure compatible with BaseLLM expectations
            return {
                "error": str(e),
                "choices": [], # Ensure choices list exists even on error
                 "model": self._model_name
            }